Remove original seed file
- `seed.ts` was not used anymore and it doesn't work with current data model
- Even if we want to `seed.ts` back, it's still in git history
- Replaced `seed.ts` with `seed2.ts`, didn't see a reason not to do it.
- Removed references to `seed2.ts/js`
- Removed matrix strategy from `check-seed-generators.yml`
