+++ test-db-manager/package.json
@@ -9,7 +9,6 @@
   "scripts": {
     "build": "yarn clean && yarn rollup --config rollup.config.mjs",
     "seed": "node dist/seed.js",
-    "seed2": "node dist/seed2.js",
     "clean": "rm -rf ts-dist && rm -rf dist && rm -f tsconfig.tsbuildinfo && rm -rf .rollup.cache",
     "ts:check": "yarn tsc --noEmit",
     "ts:build": "yarn tsc --build"
